requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
numpy==1.17.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
twilio==6.33.0
urllib3==1.25.6
pyyaml
